+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix: occur error when write file on no exist directory
 - Feat: default output file name is input file name
 - Fix: help message
+- Feat: add json output format
 
 ## Alpha